# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Astar Management API",
    "description": "API for Astar Management Platform",
    "contact": {
      "name": "Astar Works"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://local",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/editor/folders": {
      "get": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "listNodes",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "parentId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderChildrenResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "createFolder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editor/folders/{folderId}/rename": {
      "patch": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "renameFolder",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderRenameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/editor/folders/{folderId}/move": {
      "patch": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "moveFolder",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/editor/folders/{folderId}/archive": {
      "patch": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "archiveFolder",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderArchiveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editor/folders/{nodeId}/breadcrumb": {
      "get": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "getBreadcrumb",
        "parameters": [
          {
            "name": "nodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderBreadcrumbResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/editor/folders/{folderId}/children": {
      "get": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "listFolderChildren",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseFolderChildrenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/editor/folders/{folderId}": {
      "delete": {
        "tags": [
          "Editor Folders"
        ],
        "operationId": "deleteFolder",
        "parameters": [
          {
            "name": "folderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorApiResponseUnit"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EditorErrorResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "EditorNodeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "materializedPath": {
            "type": "string"
          },
          "depth": {
            "type": "integer",
            "format": "int32"
          },
          "position": {
            "type": "number",
            "format": "double"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid"
          },
          "updatedBy": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "FolderCreateRequest": {
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "maxLength": 255.0,
            "minLength": 0.0
          },
          "parentId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "number",
            "format": "double"
          }
        },
        "required": [
          "title"
        ]
      },
      "EditorApiResponseFolderResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/FolderResponse"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FolderResponse": {
        "type": "object",
        "properties": {
          "node": {
            "$ref": "#/components/schemas/EditorNodeResponse"
          }
        }
      },
...
      "FolderRenameRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255.0,
            "minLength": 0.0
          }
        },
        "required": [
          "title"
        ]
      },
      "FolderMoveRequest": {
        "type": "object",
        "properties": {
          "targetParentId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "FolderArchiveRequest": {
        "type": "object",
        "properties": {
          "archived": {
            "type": "boolean"
          }
        }
      },
...
      "EditorApiResponseFolderChildrenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/FolderChildrenResponse"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "FolderChildrenResponse": {
        "type": "object",
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EditorNodeResponse"
            }
          }
        }
      },
      "EditorApiResponseFolderBreadcrumbResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/FolderBreadcrumbResponse"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "EditorBreadcrumbItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "materializedPath": {
            "type": "string"
          }
        }
      },
      "FolderBreadcrumbResponse": {
        "type": "object",
        "properties": {
          "breadcrumb": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EditorBreadcrumbItemResponse"
            }
          }
        }
      },
...
      "EditorApiResponseUnit": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "$ref": "#/components/schemas/EditorErrorResponse"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}